% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_coordinates.R
\name{rescale_coordinates}
\alias{rescale_coordinates}
\title{Rescale x-y coordinates}
\usage{
rescale_coordinates(from, to)
}
\arguments{
\item{from}{The dimensions to convert from (see \code{help(dimensions)})}

\item{to}{The dimensions to convert to (see \code{help(dimensions)})}
}
\description{
Returns a list containing functions to rescale coordinates from
one set of pitch dimensions (i.e. data provider) to another.
}
\examples{

opta_to_wyscout <- rescale_coordinates(
  from = pitch_opta,
  to   = pitch_wyscout
)

opta_xs <- c(10, 22, 55, 78)
opta_ys <- c(10, 22, 55, 78)

opta_to_wyscout$x(opta_xs)
#> c(...)

opta_to_wyscout$y(opta_ys)
#> c(...)

}
