% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pitch.R
\name{theme_pitch}
\alias{theme_pitch}
\title{Removes background and axes details from a ggplot plot.}
\usage{
theme_pitch(force_aspect = TRUE, aspect_ratio = 68/105)
}
\arguments{
\item{force_aspect}{whether to force the plot to obey a given aspect ratio}

\item{aspect_ratio}{ratio of y / x}
}
\value{
list of ggplot themes to be added to a ggplot plot
}
\description{
Removes background and axes details from a ggplot plot.
}
\examples{
ggplot(shots_data, aes(x = x, y = y)) + pitch_layer() + geom_point() + theme_pitch()
}
