% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pitch.R
\name{goals_box}
\alias{goals_box}
\alias{goals_strip}
\alias{goals_line}
\title{Goals markings}
\usage{
goals_box(colour, fill, dimensions, offset = 2, ...)

goals_strip(
  colour,
  fill,
  dimensions,
  offset = 1,
  size = 1,
  lineend = "round",
  ...
)

goals_line(colour, fill, dimensions, ...)
}
\arguments{
\item{colour}{Colour of pitch outline.}

\item{fill}{Colour of pitch fill.}

\item{dimensions}{A list containing the pitch dimensions to draw. See \code{help(pitch_opta)}.}

\item{offset}{Determines how deep the goal extends.}

\item{...}{Passed onto underlying \code{ggplot2::annotate} calls.}

\item{size}{Determines line thickness in \code{goals_strip} and \code{goals_line}.}

\item{lineend}{Determines lineend in \code{goals_strip} and \code{goals_line}.}
}
\value{
list of ggplot geoms to be added to a ggplot plot
}
\description{
Various functions ca be supplied to \code{annotate_pitch} to specify the appearance
of goals in the resulting plot.
}
\details{
Each function takes \code{colour}, \code{fill}, and \code{dimensions} arguments. User-defined
functions with the same arguments can also be used
}
\examples{
library(ggplot2)

shots_data <- data.frame(x = c(90, 85, 82, 78, 83),
                         y = c(43, 40, 52, 56, 44))

# Default
ggplot(shots_data, aes(x = x, y = y)) +
  annotate_pitch(goals = goals_box) +
  geom_point()

# Other goals markings
ggplot(shots_data, aes(x = x, y = y)) +
  annotate_pitch(goals = goals_strip) +
  geom_point()

# Partial functions can be used to customise further
ggplot(shots_data, aes(x = x, y = y)) +
  annotate_pitch(goals = ~ goals_strip(..., size = 2)) +
  geom_point()

}
