% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch_layer.R
\name{pitch_layer}
\alias{pitch_layer}
\title{Adds soccer pitch markings as a layer for use in a ggplot plot.}
\usage{
pitch_layer(fill = "green", colour = "white", x_scale = 1, y_scale = 1,
  x_shift = 0, y_shift = 0, juego = FALSE)
}
\arguments{
\item{fill}{fill colour of the pitch}

\item{colour}{outline colour of the pitch}

\item{x_scale}{scale applied to x coordinates of the pitch markings}

\item{y_scale}{scale applied to y coordinates of the pitch markings}

\item{x_shift}{constant value added to x coordinates of the pitch markings}

\item{y_shift}{constant value added to y coordinates of the pitch markings}

\item{juego}{boolean indicating whether additional 'juego de posicion' markings should be shown}
}
\value{
list of ggplot geoms to be added to a ggplot plot
}
\description{
Adds soccer pitch markings as a layer for use in a ggplot plot.
}
\examples{
ggplot(shots_data, aes(x = x, y = y)) + pitch_layer() + geom_point() + theme_pitch()
}
