% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\docType{data}
\name{pitch_opta}
\alias{pitch_opta}
\alias{pitch_statsbomb}
\title{Pitch dimensions}
\format{An object of class \code{list} of length 10.}
\usage{
pitch_opta

pitch_statsbomb
}
\description{
Pitch dimensions
}
\keyword{datasets}
